Move to `web_app`
This code is only intended to be run in a web environment, not in Node.
