+++ src/server_manager/web_app/data_formatting.ts
